hihi

+++ CaseStudy/app/Http/Controllers/CategoriesController.php
@@ -2,23 +2,22 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Category;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
-use App\Models\Category;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Storage;
 
 class CategoriesController extends Controller
 {
-    public function home()
-    {
-        return view('index');
-    }
+   
     public function index()
     {
         $this->authorize('viewAny', Category::class);
-        $items = Category::all();
+        $items = Category::search()->paginate(3);
         return view('admin.categories.index', compact('items'));
+
     }
     public function create()
     {
@@ -28,6 +27,7 @@ class CategoriesController extends Controller
     }
     public function store(StoreCategoryRequest $request)
     {
+        $this->authorize('create', Category::class);
         $category = new Category();
         $category->name = $request->name;
         $fieldName = 'inputFile';
@@ -45,19 +45,27 @@ class CategoriesController extends Controller
             toast(__('messages.msg_cate_add_ss', ['name' => $request->name]), 'success', 'top-right');
             return redirect()->route('categories');
         } catch (\Exception$e) {
+            if(isset($path)){
+
+                $images = str_replace('storage', 'public', $path);
+                Storage::delete($images);
+            }
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
+
             toast(__('messages.msg_cate_add_err', ['name' => $request->name]), 'error', 'top-right');
             return view('admin.categories.add', compact('request'));
         }
     }
     public function edit($id)
     {
-        $item = Category::find($id);
         $this->authorize('update', Category::class);
+        $item = Category::find($id);
         return view('admin.categories.edit', compact('item'));
     }
     public function update(UpdateCategoryRequest $request, $id)
     {
-        $category = Category::find($id);
+        $this->authorize('update', Category::class);
+        $category = Category::findOrFail($id);
         $category->name = $request->name;
         $fieldName = 'inputFile';
         if ($request->hasFile($fieldName)) {
@@ -74,37 +82,38 @@ class CategoriesController extends Controller
             $images = str_replace('storage', 'public', $item->image);
         }
         try {
-            $item->save();
-            if (isset($path)) {
+            $category->save();
+            if (isset($path) && isset($images)) {
                 Storage::delete($images);
             }
-            $item->delete();
-            return response()->json([
-                'code' => 200,
-                'message' => 'success',
-            ], status:200);
+            toast(__('messages.msg_cate_up_ss', ['name' => $request->name]), 'success', 'top-right');
+            return redirect()->route('categories');
         } catch (\Exception$e) {
-            $images = $images = str_replace('storage', 'public', $path);
-            Storage::delete($images);
-            $item->delete();
-            return response()->json([
-                'code' => 201,
-                'message' => 'success',
-            ], status:200);
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
+
+            if (isset($path)) {
+                $images = str_replace('storage', 'public', $path);
+                Storage::delete($images);
+            }
+            toast(__('messages.msg_cate_up_err', ['name' => $request->name]), 'error', 'top-right');
+            return redirect()->route('categories.edit', $category->id);
         }
+        return redirect()->route('categories');
     }
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        $item = Category::findOrFail($id);
         $this->authorize('delete', Category::class);
+        $id = $request->id;
+        $item = Category::findOrFail($id);
         try {
             $item->delete();
             return response()->json([
                 'code' => 200,
                 'message' => 'success',
             ], status:200);
-   
+
         } catch (\Exception$e) {
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
             return response()->json([
                 'code' => 201,
                 'message' => 'success',
@@ -118,15 +127,16 @@ class CategoriesController extends Controller
     }
     public function garbageCan()
     {
-        $items = Category::onlyTrashed()->paginate(5);
+        $this->authorize('viewgc', Category::class);
+        $items = Category::search()->onlyTrashed()->paginate(3);
         return view('admin.categories.Garbage_can', compact('items'));
 
     }
     public function restore($id)
     {
         try {
-            $item = Category::withTrashed()->where('id', $id);
             $this->authorize('restore', Category::class);
+            $item = Category::withTrashed()->where('id', $id);
             $item->restore();
             $item = Category::findOrFail($id);
             return response()->json([
@@ -134,6 +144,7 @@ class CategoriesController extends Controller
                 'message' => 'success',
             ], status:200);
         } catch (\Exception$e) {
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
             return response()->json([
                 'code' => 404,
                 'message' => 'failed',
@@ -142,24 +153,22 @@ class CategoriesController extends Controller
     }
     public function forceDelete($id)
     {
-        DB::beginTransaction();
-        $item = Category::onlyTrashed()->findOrFail($id);
         $this->authorize('forceDelete', Category::class);
+        $item = Category::onlyTrashed()->findOrFail($id);
         $images = str_replace('storage', 'public', $item->image);
-        $items = Category::withTrashed()->where('id', $id)->forceDelete();
         try {
+            Category::withTrashed()->where('id', $id)->forceDelete();
             Storage::delete($images);
-            DB::commit();
             return response()->json([
                 'code' => 200,
                 'message' => 'success',
             ], status:200);
         } catch (\Exception$e) {
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
             return response()->json([
-                'code' => 404,
-                'message' => 'failed',
+                'code' => 201,
+                'message' => 'error',
             ], status:200);
-            DB::rollBack();
         }
     }
 }
