+++ CaseStudy/app/Http/Controllers/ProductsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\Category;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
@@ -14,8 +15,17 @@ class ProductsController extends Controller
     public function index()
     {
         $this->authorize('viewAny', Product::class);
-        $items = Product::paginate(5);
+        $items = Product::search()->paginate(3);
         return view('admin.products.index', compact('items'));
+    }
+    public function productOutOfStock()
+    {
+        $this->authorize('viewAny', Product::class);
+        $items = Product::search()->where('quantity','<', 1)->paginate(3);
+        $params = [
+            'items' => $items,
+        ];
+        return view('admin.products.outofstock',$params);
     }
     public function create()
     {   $this->authorize('create', Product::class);
@@ -29,6 +39,7 @@ class ProductsController extends Controller
     }
     public function store(StoreProductRequest $request)
     {
+        $this->authorize('create', Product::class);
         $categories = Category::all();
         $items = Product::all();
         $params = [
@@ -56,7 +67,6 @@ class ProductsController extends Controller
 
         $products->color = $request->color;
         $products->price_product = $request->price_product;
-        $products->garbage_can = 1;
         $products->category_id = $request->category_id;
         $products->user_id = Auth()->user()->id;
         try {
@@ -66,24 +76,26 @@ class ProductsController extends Controller
             toast(__('messages.msg_prd_add_ss',['name' => $request->name]),'success','top-right');
             return redirect()->route('products');
         } catch (\Exception$e) {
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
             $images = str_replace('storage', 'public', $path);
             Storage::delete($images);
             // alert()->error('Thêm Sản Phẩm: ' . $request->name, 'Không Thành Công!');
-            toast(__('messages.msg_prd_add_err',['name' => $request->name]),'error','top-right');
+            // toast(__('messages.msg_prd_add_err',['name' => $request->name]),'error','top-right');
             // return redirect()->route('products');
             return view('admin.products.add', $params);
         }
     }
     public function edit($id)
     {
+        $this->authorize('update', Product::class);
         $item = Product::find($id);
-        $this->authorize('update', $item);
         $categories = Category::all();
         return view('admin.products.edit', compact('item', 'categories'));
     }
     public function update(UpdateProductRequest $request, $id)
     {
-        $products = Product::find($id);
+        $this->authorize('update', Product::class);
+        $products = Product::findOrFail($id);
         $products->name = $request->name;
         $products->price = $request->price;
         $products->describe = $request->describe;
@@ -106,23 +118,20 @@ class ProductsController extends Controller
         }
         $products->color = $request->color;
         $products->price_product = $request->price_product;
-        $products->garbage_can = 1;
         $products->category_id = $request->category_id;
-        // Session::flash('success', 'Chỉnh sửa thành công '.$request->name);
         try {
             $products->save();
-            if(isset($path)){
+            if(isset($path) && isset($images)){
                 Storage::delete($images);
             }
-            // alert()->success('Lưu Sản Phẩm: ' . $request->name, 'Thành Công');
             toast(__('messages.msg_prd_up_ss',['name' => $request->name]),'success','top-right');
             return redirect()->route('products');
-
         } catch (\Exception $e) {
-            // $products = Product::find($id);
-            $images = $images = str_replace('storage', 'public', $path);
-            Storage::delete($images);
-            // alert()->error('Lưu Sản Phẩm: ' . $request->name, 'Không Thành Công!');
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
+            if(isset($path)){
+                $images = str_replace('storage', 'public', $path);
+                Storage::delete($images);
+            }
             toast(__('messages.msg_prd_up_err',['name' => $request->name]),'error','top-right');
             return redirect()->route('products.edit',$products->id);
         }
@@ -130,61 +139,70 @@ class ProductsController extends Controller
     }
     public function destroy($id)
     {
+        $this->authorize('delete', Product::class);
         $item = Product::findOrFail($id);
-        $this->authorize('delete', $item);
         try {
             $item->delete();
-            // alert()->success('Xóa Sản Phẩm: ' . $item->name, 'Thành Công');
-            toast(__('messages.msg_prd_dele_ss',['name' => $item->name]),'success','top-right');
-            return redirect()->route('products');
-
+            return response()->json([
+                'code' => 200,
+                'message' => 'success',
+            ], status:200);
         } catch (\Exception$e) {
-            // alert()->error('Xóa Sản Phẩm: ' . $item->name, 'Không Thành Công!');
-            toast(__('messages.msg_prd_dele_err',['name' => $item->name]),'error','top-right');
-            return redirect()->route('products');
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
+            return response()->json([
+                'code' => 201,
+                'message' => 'error',
+            ], status:200);
         }
-        // Session::flash('success', 'Xóa thành công '.$item->name);
     }
-
     public function garbageCan(){
-        $items = Product::onlyTrashed()->paginate(5);
+        $this->authorize('viewgc', Product::class);
+        $items = Product::search()->onlyTrashed()->paginate(5);
         return view('admin.products.Garbage_can', compact('items'));
      
     }
     public function restore($id){
         try {
+            $this->authorize('restore', Product::class);
             $item = Product::withTrashed()->where('id', $id);
-            $this->authorize('restore', $item);
             $item->restore();
             $item = Product::findOrFail($id);
-            alert()->success('Khôi Phục Sản Phẩm: ' . $item->name, 'Thành Công');
+            return response()->json([
+                'code' => 200,
+                'message' => 'success',
+            ], status:200);
             return redirect()->route('products.garbageCan');
         } catch (\Exception$e) {
-            alert()->error('Khôi Phục Sản Phẩm: ' . $item->name, 'Không Thành Công!');
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
+            return response()->json([
+                'code' => 201,
+                'message' => 'error',
+            ], status:200);
             return redirect()->route('products.garbageCan');
         }
-        //Xoá record vĩnh viễn: App\User::withTrashed()->where('id', 1)->forceDelete();
-        //Để lấy lại record đã xoá bằng softDeletes: App\User::withTrashed()->where('id', 1)->restore();
     }
     public function forceDelete($id){
+        $this->authorize('forceDelete', Product::class);
         DB::beginTransaction();
         $item = Product::onlyTrashed()->findOrFail($id);
-        $this->authorize('forceDelete', $item);
-        // dd($product);
         $images = str_replace('storage', 'public', $item->image);
       
         $item = Product::withTrashed()->where('id', $id)->forceDelete();
         try {
-            // alert()->success('Xóa Sản Phẩm: ' . $product->name, 'Thành Công');
-            toast(__('messages.msg_prd_dele_ss',['name' => $item->name]),'success','top-right');
             Storage::delete($images);
             DB::commit();
+            return response()->json([
+                'code' => 200,
+                'message' => 'success',
+            ], status:200);
             return redirect()->route('products.garbageCan');
         } catch (\Exception$e) {
-            // alert()->error('Xóa Sản Phẩm: ' . $product->name, 'Không Thành Công!');
-            toast(__('messages.msg_prd_dele_err',['name' => $item->name]),'error','top-right');
+               Log::error('message: ' . $e->getMessage() . ' line: ' . $e->getLine() . ' file: ' . $e->getFile());
             DB::rollBack();
-            return redirect()->route('products.garbageCan');
+            return response()->json([
+                'code' => 201,
+                'message' => 'error',
+            ], status:200);
         }
     }
